Update __init__.py
Changed the way that the program was generating a key. It now uses an actual list instead of a string. The string was what caused some of the type errors before, I was like a fool attempting to index through the string using a integer. This revisions make_alpha function can edit a alphabet(saved as a list) by shifting over the location of all the characters in the array.
    Might redit the program so that when 0 is given as an argument for start_point it throws an error instead of just derping.